Restore a working linking setup
The purpose of this is to restore a working setup where we can link
all glimmer-vm to ember.js without publishing.

Previously this was just a script that `yarn link` on both sides.

It seems like we tried to use `yalc` at some point.

Neither of those other solutions fully works with the constraints:

* supports pnpm
* supports pnpm workspace
* supports `publishConfig`
* resolve internal dependencies (like `@glimmer/encoder`) when
  linked

This script makes it work with all of the above.

Usage:

1. pnpm install, develop, make changes, etc
2. pnpm build
3. pnpm link:all
4. pnpm link:glimmer-vm from ember.js
5. validate, iterate, re-run pnpm build/pnpm link:all as needed
6. pnpm unlink from ember.js
7. pnpm unlink:all here

See also https://github.com/emberjs/ember.js/pull/20590