Revert some VSCode settings
In this particular case, it was the setting to hide ignored files
in the explorer. It is often useful to go into those folders (like
`node_modules` and `dist`) when troubleshooting, working on the
build, trying to patch a package, etc.

In any case, if that is the desired behavior one can always add it
to their global user settings, I don't think there is something
bespoke about this repo that requires this setting any more so than
other workspaces.

In general, the principle is probably things like these should stay
out of the committed configs and belongs to the users's settings,
and there are probably other settings like that which should be
backed out as well, this is just the one I happen to notice and I
don't have much time to go through the rest.