Improved error messages and added validation util
Syntax errors are now notes in the AST rather than reported during parse
time, and are reported at compile-time. To make it possible to get
syntax errors purely from `@glimmer/syntax`, this commit adds a verifier
to `@glimmer/syntax` that validates the standalone `ASTv2` and reports
errors that it finds.

To get a complete picture, you really _do_ want to wait until
compile-time, which can report keyword-specific errors, but this makes
it possible to get a sense of the syntax errors without involving the
compiler.

This is all in service of making errors reported by `@glimmer/syntax`
fully forgiving.